fix adc

+++ User_Hardware/u_adc.h
@@ -1,7 +1,7 @@
 #ifndef __U_ADC_H__
 #define __U_ADC_H__
 
-extern volatile unsigned int adc_value[4][5];
+extern volatile float adc_value[4][5];
 
 void U_ADC_Init(void);
 void U_ADC_Read(void);
